refactor: Modularize AIChat page with separate components
